Give the fireball more precision when fired

+++ include/CGame.h
@@ -12,7 +12,7 @@ namespace odb {
         std::array<std::array< int, 40>, 40 > mMap;
         std::vector<std::shared_ptr<CActor>> mActors;
         RayCollision castRay(int offset);
-        void spawnFireball( int x, int y, float angle );
+        void spawnFireball( float x, float y, float angle );
     public:
         explicit CGame();
         void tick( long ms );
